# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StoryBloom API – AI children's story generation and management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Story {
  id: number;
  childName: string;
  childAge: number;
  childGender: string;
  /** @nullable */
  milestones?: string | null;
  theme: string;
  /** @nullable */
  customPrompt?: string | null;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export type StoryInputChildGender = typeof StoryInputChildGender[keyof typeof StoryInputChildGender];


export const StoryInputChildGender = {
  boy: 'boy',
  girl: 'girl',
  neutral: 'neutral',
} as const;

export type StoryInputTheme = typeof StoryInputTheme[keyof typeof StoryInputTheme];


export const StoryInputTheme = {
  Courage: 'Courage',
  Kindness: 'Kindness',
  Overcoming_Fear: 'Overcoming Fear',
  Faith: 'Faith',
  Friendship: 'Friendship',
  Honesty: 'Honesty',
  Perseverance: 'Perseverance',
  Gratitude: 'Gratitude',
} as const;

export interface StoryInput {
  /** @minLength 1 */
  childName: string;
  /**
     * @minimum 1
     * @maximum 12
     */
  childAge: number;
  childGender: StoryInputChildGender;
  /** Details like favorite toys, pets, recent trips */
  milestones?: string;
  theme: StoryInputTheme;
  /** Custom topic or plot prompt */
  customPrompt?: string;
}

export interface StoryUpdate {
  title?: string;
  content?: string;
  childName?: string;
  childAge?: number;
  childGender?: string;
  milestones?: string;
  theme?: string;
  customPrompt?: string;
}

export interface ThemeCount {
  theme: string;
  count: number;
}

export interface StoryStats {
  totalStories: number;
  byTheme: ThemeCount[];
  /** @nullable */
  mostRecentChildName: string | null;
}