# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StoryBloom API – AI children's story generation and management
 * OpenAPI spec version: 0.1.0
 */

export interface Story {
  id: number;
  childName: string;
  childAge: number;
  childGender: string;
  /** @nullable */
  milestones?: string | null;
  theme: string;
  /** @nullable */
  customPrompt?: string | null;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}