# lib/api-zod/src/generated/types/storyInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StoryBloom API – AI children's story generation and management
 * OpenAPI spec version: 0.1.0
 */
import type { StoryInputChildGender } from './storyInputChildGender';
import type { StoryInputTheme } from './storyInputTheme';

export interface StoryInput {
  /** @minLength 1 */
  childName: string;
  /**
     * @minimum 1
     * @maximum 12
     */
  childAge: number;
  childGender: StoryInputChildGender;
  /** Details like favorite toys, pets, recent trips */
  milestones?: string;
  theme: StoryInputTheme;
  /** Custom topic or plot prompt */
  customPrompt?: string;
}