# lib/api-zod/src/generated/types/storyUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StoryBloom API – AI children's story generation and management
 * OpenAPI spec version: 0.1.0
 */

export interface StoryUpdate {
  title?: string;
  content?: string;
  childName?: string;
  childAge?: number;
  childGender?: string;
  milestones?: string;
  theme?: string;
  customPrompt?: string;
}